seo-automation/requirements.txt
anthropic>=0.80.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
google-auth>=2.29.0
google-api-python-client>=2.120.0
python-dotenv>=1.0.0
